deploy/pyproject.toml
[tool.poetry]
name = "equinor-trading-repo-template"
version = "0.1.0"
description = "Standard template containing the essentials for a new repo"
authors = ["Equinor"]

[tool.poetry.dependencies]
python = ">=3.7.1,<3.9"
pulumi = "^3.19.0"
pulumi-azure-native = "^1.48.0"
ipykernel = "^6.6.0"
pulumi-docker = "^3.1.0"
requests = "^2.26.0"
azure-identity = "^1.7.1"
msgraph-core = "^0.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
